+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Fixed
+* Issue where calling load on `AnchorList` with default anchor types would potentially
+    add duplicate anchor types
+
+
+
+
 ## [0.7.0] - 2023-08-25
 
 ### Added